misc/HtmlTestSuite/results/test-0107.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
    <script>
      setup({explicit_done: true})
      window.addEventListener("message", e =&gt; {
        test(() =&gt; {
          assert_equals(e.data, "PASS", "The iframe saw ServerTiming entries in its NavigationTiming entry.");
          done();
        })
      });
    </script>
</head>
<body>
<iframe src="resources/navigation-timing.html?pipe=trickle(800:d1)"/>
</body>


</html>
